Introduce separate suite for multi-session
